api-rasa/requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
bidict==0.22.1
blinker==1.7.0
blis==0.7.11
boto3==1.34.7
botocore==1.34.7
build==0.9.0
CacheControl==0.12.14
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==1.9.2
cryptography==3.4.6
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dnspython==1.16.0
docopt==0.6.2
fasttext-wheel==0.9.2
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.12
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.5.3
mattermostwrapper==2.2
msgpack==1.0.7
multidict==5.2.0
multipart==0.2.4
murmurhash==1.0.10
networkx==2.6.3
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pamqp==3.2.1
pandas==2.1.4
partd==1.4.1
pdfservices-sdk==2.3.0
pep517==0.13.0
Pillow==10.1.0
pluggy==1.3.0
polling==0.3.2
polling2==0.5.0
preshed==3.0.9
prompt-toolkit==3.0.28
proto-plus==1.23.0
protobuf==3.20.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
Pygments==2.14.0
PyJWT==2.4.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2022.7.1
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.5.4
rasa-sdk==3.5.0
redis==4.6.0
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
safetensors==0.4.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.26.1
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.50
srsly==2.4.8
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
transformers==4.36.2
twilio==7.14.2
typeguard==4.1.5
typer==0.9.0
typing-utils==0.1.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
underthesea==6.8.0
underthesea_core==1.0.4
urllib3==1.26.13
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
